move tokens
