+++ src/swap/hooks/useSwapLiteTokens.ts
@@ -2,7 +2,7 @@ import { useState } from 'react';
 import type { Address } from 'viem';
 import { useValue } from '../../core-react/internal/hooks/useValue';
 import type { Token } from '../../token';
-import { ethToken, usdcToken } from '../constants';
+import { ethToken, usdcToken } from '../../token/constants';
 import type { SwapLiteTokens } from '../types';
 import { useSwapBalances } from './useSwapBalances';
